Move GenericNotification into `components/` to give the hint it's an internal implementation
