+++ components/Notification/components/GenericNotification.module.scss
@@ -1,4 +1,4 @@
-@import 'styles';
+@import '../styles';
 
 .notification {
   @extend %ca-notification;
