+++ components/Notification/index.js
@@ -1,12 +1,6 @@
 // @flow
-import GenericNotification from './GenericNotification.js';
 import GlobalNotification from './GlobalNotification.js';
 import InlineNotification from './InlineNotification.js';
 import ToastNotification from './ToastNotification.js';
 
-export {
-  GenericNotification,
-  GlobalNotification,
-  InlineNotification,
-  ToastNotification,
-};
+export { GlobalNotification, InlineNotification, ToastNotification };
